feat(plugin-cloud-storage): sanitize prefixes in getFileKey and simplify test configs
- Sanitize collection and doc segments in getFileKey for storage object keys
- Pass options.prefix through plugin and admin without sanitizePrefix or empty-string fallback
- Match admin prefix field defaultValue to server getFields
- Add buildPluginCloudStorageIntConfig with config.ts vs config.compositePrefixes.ts
- Split composite S3 tests into int.compositePrefixes.int.spec.ts; tighten int.spec types
