+++ test/plugin-cloud-storage/config.ts
@@ -1,230 +1,4 @@
-import type { S3StorageOptions } from '@payloadcms/storage-s3'
-import type { Plugin } from 'payload'
+import { buildPluginCloudStorageIntConfig } from './buildPluginCloudStorageIntConfig.js'
 
-import { cloudStoragePlugin } from '@payloadcms/plugin-cloud-storage'
-import { azureStorage } from '@payloadcms/storage-azure'
-import { gcsStorage } from '@payloadcms/storage-gcs'
-import { s3Storage } from '@payloadcms/storage-s3'
-import dotenv from 'dotenv'
-import { fileURLToPath } from 'node:url'
-import path from 'path'
-
-import { buildConfigWithDefaults } from '../buildConfigWithDefaults.js'
-import { devUser } from '../credentials.js'
-import { Media } from './collections/Media.js'
-import { MediaWithCompositePrefixes } from './collections/MediaWithCompositePrefixes.js'
-import { MediaWithCustomURL } from './collections/MediaWithCustomURL.js'
-import { MediaWithGenerateFileURL } from './collections/MediaWithGenerateFileURL.js'
-import { MediaWithPrefix } from './collections/MediaWithPrefix.js'
-import { RestrictedMedia } from './collections/RestrictedMedia.js'
-import { TestMetadata } from './collections/TestMetadata.js'
-import { Users } from './collections/Users.js'
-import {
-  collectionPrefix,
-  mediaSlug,
-  mediaWithCompositePrefixesSlug,
-  mediaWithCustomURLSlug,
-  mediaWithGenerateFileURLSlug,
-  mediaWithPrefixSlug,
-  prefix,
-  restrictedMediaSlug,
-  testMetadataSlug,
-} from './shared.js'
-import { createTestBucket } from './utils.js'
-
-const filename = fileURLToPath(import.meta.url)
-const dirname = path.dirname(filename)
-
-let storagePlugin: Plugin = {} as Plugin
-let uploadOptions
-
-// Load config to work with emulated services
-dotenv.config({
-  path: path.resolve(dirname, './.env.emulated'),
-})
-
-if (process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER === 'azure') {
-  storagePlugin = azureStorage({
-    collections: {
-      [mediaSlug]: true,
-      [mediaWithPrefixSlug]: {
-        prefix,
-      },
-      [restrictedMediaSlug]: true,
-    },
-    allowContainerCreate: process.env.AZURE_STORAGE_ALLOW_CONTAINER_CREATE === 'true',
-    baseURL: process.env.AZURE_STORAGE_ACCOUNT_BASEURL!,
-    connectionString: process.env.AZURE_STORAGE_CONNECTION_STRING!,
-    containerName: process.env.AZURE_STORAGE_CONTAINER_NAME!,
-  })
-}
-
-if (process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER === 'gcs') {
-  storagePlugin = gcsStorage({
-    collections: {
-      [mediaSlug]: true,
-      [mediaWithPrefixSlug]: {
-        prefix,
-      },
-      [restrictedMediaSlug]: true,
-    },
-    bucket: process.env.GCS_BUCKET!,
-    options: {
-      apiEndpoint: process.env.GCS_ENDPOINT,
-      projectId: process.env.GCS_PROJECT_ID,
-    },
-  })
-}
-
-if (
-  process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER === 's3' ||
-  !process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER
-) {
-  // The s3 adapter supports using temp files for uploads
-  uploadOptions = {
-    useTempFiles: true,
-  }
-
-  storagePlugin = s3Storage({
-    collections: {
-      [mediaSlug]: true,
-      [mediaWithPrefixSlug]: {
-        prefix,
-      },
-      [mediaWithCompositePrefixesSlug]: {
-        prefix: collectionPrefix,
-      },
-      [mediaWithCustomURLSlug]: {
-        prefix,
-        disablePayloadAccessControl: true,
-        generateFileURL: ({ filename, prefix }) =>
-          filename
-            ? `https://test-cdn.example.com/${prefix}/${encodeURIComponent(filename)}`
-            : null,
-      } as S3StorageOptions['collections'][keyof S3StorageOptions['collections']],
-      [mediaWithGenerateFileURLSlug]: {
-        prefix,
-        // disablePayloadAccessControl defaults to false - files proxied through Payload
-        // This tests that generateFileURL works even when Payload proxies files
-        generateFileURL: ({ filename, prefix }) =>
-          filename
-            ? `https://cdn-proxied.example.com/${prefix}/${encodeURIComponent(filename)}`
-            : null,
-      } as S3StorageOptions['collections'][keyof S3StorageOptions['collections']],
-      [restrictedMediaSlug]: true,
-    },
-    bucket: process.env.S3_BUCKET ?? '',
-    useCompositePrefixes: true,
-    config: {
-      credentials: {
-        accessKeyId: process.env.S3_ACCESS_KEY_ID ?? '',
-        secretAccessKey: process.env.S3_SECRET_ACCESS_KEY ?? '',
-      },
-      endpoint: process.env.S3_ENDPOINT,
-      forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
-      region: process.env.S3_REGION,
-    },
-  })
-}
-
-if (process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER === 'r2') {
-  s3Storage({
-    collections: {
-      [mediaSlug]: true,
-      [mediaWithPrefixSlug]: {
-        prefix,
-      },
-    },
-    bucket: process.env.R2_BUCKET ?? '',
-    config: {
-      credentials: {
-        accessKeyId: process.env.S3_ACCESS_KEY_ID ?? '',
-        secretAccessKey: process.env.S3_SECRET_ACCESS_KEY ?? '',
-      },
-      endpoint: process.env.S3_ENDPOINT,
-      forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
-      region: process.env.S3_REGION,
-    },
-  })
-}
-
-const testMetadataPlugin = cloudStoragePlugin({
-  collections: {
-    [testMetadataSlug]: {
-      adapter: () => ({
-        name: 'test-metadata-adapter',
-        handleUpload: ({ file, data }) => {
-          const metadata = {
-            ...data,
-            customStorageId: `storage-${Date.now()}`,
-            uploadTimestamp: new Date().toISOString(),
-            storageProvider: 'test-adapter',
-            bucketName: 'test-bucket',
-            objectKey: data.filename || file.filename, // Use the processed filename from data
-            processingStatus: 'completed',
-            uploadVersion: '1.0.0',
-          }
-          return metadata
-        },
-        handleDelete: () => Promise.resolve(),
-        staticHandler: () => new Response('Not found', { status: 404 }),
-      }),
-    },
-  },
-})
-
-export default buildConfigWithDefaults({
-  admin: {
-    importMap: {
-      baseDir: path.resolve(dirname),
-    },
-  },
-  collections: [
-    Media,
-    MediaWithCompositePrefixes,
-    MediaWithCustomURL,
-    MediaWithGenerateFileURL,
-    MediaWithPrefix,
-    RestrictedMedia,
-    TestMetadata,
-    Users,
-  ],
-  onInit: async (payload) => {
-    /*const client = new AWS.S3({
-      endpoint: process.env.S3_ENDPOINT,
-      forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
-      region: process.env.S3_REGION,
-      credentials: {
-        accessKeyId: process.env.S3_ACCESS_KEY_ID,
-        secretAccessKey: process.env.S3_SECRET_ACCESS_KEY,
-      },
-    })
-
-    const makeBucketRes = await client.send(
-      new AWS.CreateBucketCommand({ Bucket: 'payload-bucket' }),
-    )
-
-    if (makeBucketRes.$metadata.httpStatusCode !== 200) {
-      throw new Error(`Failed to create bucket. ${makeBucketRes.$metadata.httpStatusCode}`)
-    }*/
-
-    await payload.create({
-      collection: 'users',
-      data: {
-        email: devUser.email,
-        password: devUser.password,
-      },
-    })
-
-    await createTestBucket()
-
-    payload.logger.info(
-      `Using plugin-cloud-storage adapter: ${process.env.PAYLOAD_PUBLIC_CLOUD_STORAGE_ADAPTER}`,
-    )
-  },
-  plugins: [storagePlugin, testMetadataPlugin],
-  upload: uploadOptions,
-  typescript: {
-    outputFile: path.resolve(dirname, 'payload-types.ts'),
-  },
-})
+/** Default integration config: S3 `useCompositePrefixes: false` (non-composite prefix + URL behavior). */
+export default buildPluginCloudStorageIntConfig({ useCompositePrefixes: false })
