+++ test/plugin-cloud-storage/int.spec.ts
@@ -14,9 +14,7 @@ import type { Config } from './payload-types.js'
 
 import { initPayloadInt } from '../__helpers/shared/initPayloadInt.js'
 import {
-  collectionPrefix,
   mediaSlug,
-  mediaWithCompositePrefixesSlug,
   mediaWithCustomURLSlug,
   mediaWithGenerateFileURLSlug,
   mediaWithPrefixSlug,
@@ -29,7 +27,57 @@ import { clearTestBucket, createTestBucket } from './utils.js'
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
-let payload: Payload
+async function verifyUploads({
+  client,
+  collectionSlug,
+  payload,
+  prefix = '',
+  TEST_BUCKET,
+  uploadId,
+}: {
+  client: AWS.S3Client
+  collectionSlug: keyof Config['collections']
+  payload: Payload
+  prefix?: string
+  TEST_BUCKET: string
+  uploadId: number | string
+}) {
+  const uploadData = (await payload.findByID({
+    id: uploadId,
+    collection: collectionSlug,
+  })) as unknown as { filename: string; sizes: Record<string, { filename: string }> }
+
+  const sizes = uploadData.sizes ?? {}
+  const fileKeys = Object.keys(sizes).map((key) => {
+    const entry = sizes[key]
+    if (!entry) {
+      throw new Error(`Missing sizes entry for key: ${key}`)
+    }
+
+    const rawFilename = entry.filename
+    return prefix ? `${prefix}/${rawFilename}` : rawFilename
+  })
+
+  fileKeys.push(`${prefix ? `${prefix}/` : ''}${uploadData.filename}`)
+  try {
+    for (const key of fileKeys) {
+      const { $metadata } = await client.send(
+        new AWS.HeadObjectCommand({ Bucket: TEST_BUCKET, Key: key }),
+      )
+
+      if ($metadata.httpStatusCode !== 200) {
+        console.error('Error verifying uploads', key, $metadata)
+        throw new Error(`Error verifying uploads: ${key}, ${$metadata.httpStatusCode}`)
+      }
+
+      // Verify each size was properly uploaded
+      expect($metadata.httpStatusCode).toBe(200)
+    }
+  } catch (error: unknown) {
+    console.error('Error verifying uploads:', fileKeys, error)
+    throw error
+  }
+}
 
 // needs to be here as it imports vitest functions and conflicts with playwright that uses jest
 export function describeIfInCIOrHasLocalstack(): SuiteAPI | SuiteAPI['skip'] {
@@ -51,17 +99,6 @@ export function describeIfInCIOrHasLocalstack(): SuiteAPI | SuiteAPI['skip'] {
 }
 
 describe('@payloadcms/plugin-cloud-storage', () => {
-  let TEST_BUCKET: string
-
-  beforeAll(async () => {
-    ;({ payload } = await initPayloadInt(dirname))
-    TEST_BUCKET = process.env.S3_BUCKET
-  })
-
-  afterAll(async () => {
-    await payload.destroy()
-  })
-
   describe('getFilePrefix', () => {
     const mockReq = {
       payload: {
@@ -226,457 +263,397 @@ describe('@payloadcms/plugin-cloud-storage', () => {
     })
   })
 
-  let client: AWS.S3Client
-  describeIfInCIOrHasLocalstack()('plugin-cloud-storage', () => {
-    describe('S3', () => {
-      beforeAll(async () => {
-        client = new AWS.S3({
-          endpoint: process.env.S3_ENDPOINT,
-          forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
-          region: process.env.S3_REGION,
-          credentials: {
-            accessKeyId: process.env.S3_ACCESS_KEY_ID,
-            secretAccessKey: process.env.S3_SECRET_ACCESS_KEY,
-          },
+  describe('integration (non-composite prefixes)', () => {
+    let payload: Payload
+    let TEST_BUCKET: string
+
+    beforeAll(async () => {
+      ;({ payload } = await initPayloadInt(dirname, undefined, true, 'config.ts'))
+      TEST_BUCKET = process.env.S3_BUCKET!
+    })
+
+    afterAll(async () => {
+      await payload.destroy()
+    })
+
+    let client: AWS.S3Client
+    describeIfInCIOrHasLocalstack()('plugin-cloud-storage (non-composite prefixes)', () => {
+      describe('S3', () => {
+        beforeAll(async () => {
+          client = new AWS.S3({
+            credentials: {
+              accessKeyId: process.env.S3_ACCESS_KEY_ID!,
+              secretAccessKey: process.env.S3_SECRET_ACCESS_KEY!,
+            },
+            endpoint: process.env.S3_ENDPOINT!,
+            forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
+            region: process.env.S3_REGION!,
+          })
+
+          await createTestBucket()
+          await clearTestBucket(client)
         })
 
-        await createTestBucket()
-        await clearTestBucket(client)
-      })
-
-      afterEach(async () => {
-        await clearTestBucket(client)
-      })
-
-      it('can upload', async () => {
-        const upload = await payload.create({
-          collection: mediaSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
+        afterEach(async () => {
+          await clearTestBucket(client)
         })
 
-        expect(upload.id).toBeTruthy()
+        it('can upload', async () => {
+          const upload = await payload.create({
+            collection: mediaSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
 
-        await verifyUploads({
-          collectionSlug: mediaSlug,
-          uploadId: upload.id,
+          expect(upload.id).toBeTruthy()
+
+          await verifyUploads({
+            client,
+            collectionSlug: mediaSlug,
+            payload,
+            TEST_BUCKET,
+            uploadId: upload.id,
+          })
+
+          expect(upload.url).toEqual(`/api/${mediaSlug}/file/${String(upload.filename)}`)
         })
 
-        expect(upload.url).toEqual(`/api/${mediaSlug}/file/${String(upload.filename)}`)
-      })
+        it('can upload with prefix', async () => {
+          const upload = await payload.create({
+            collection: mediaWithPrefixSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
 
-      it('can upload with prefix', async () => {
-        const upload = await payload.create({
-          collection: mediaWithPrefixSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
+          expect(upload.id).toBeTruthy()
+
+          await verifyUploads({
+            client,
+            collectionSlug: mediaWithPrefixSlug,
+            payload,
+            prefix,
+            TEST_BUCKET,
+            uploadId: upload.id,
+          })
+          expect(upload.url).toEqual(
+            `/api/${mediaWithPrefixSlug}/file/${String(upload.filename)}?prefix=${prefix}`,
+          )
         })
 
-        expect(upload.id).toBeTruthy()
+        it('should not upload to S3 when mimeType validation fails', async () => {
+          // Get initial bucket contents
+          const listBefore = await client.send(
+            new AWS.ListObjectsV2Command({ Bucket: TEST_BUCKET }),
+          )
+          const fileCountBefore = listBefore.Contents?.length || 0
 
-        await verifyUploads({
-          collectionSlug: mediaWithPrefixSlug,
-          uploadId: upload.id,
-          prefix,
-        })
-        expect(upload.url).toEqual(
-          `/api/${mediaWithPrefixSlug}/file/${String(upload.filename)}?prefix=${prefix}`,
-        )
-      })
+          // Try to upload a JSON file to a collection that only accepts PNG
+          await expect(
+            payload.create({
+              collection: restrictedMediaSlug,
+              data: {},
+              filePath: path.resolve(dirname, './test.json'),
+            }),
+          ).rejects.toThrow()
 
-      it('can upload with composite prefixes (collection + doc prefix)', async () => {
-        const docPrefix = 'user-123'
+          // Verify no new files were uploaded to S3
+          const listAfter = await client.send(new AWS.ListObjectsV2Command({ Bucket: TEST_BUCKET }))
+          const fileCountAfter = listAfter.Contents?.length || 0
 
-        const upload = await payload.create({
-          collection: mediaWithCompositePrefixesSlug,
-          data: {
-            prefix: docPrefix,
-          },
-          filePath: path.resolve(dirname, '../uploads/image.png'),
-        })
-
-        expect(upload.id).toBeTruthy()
-
-        // With useCompositePrefixes: true, the file should be at collectionPrefix/docPrefix/filename
-        const expectedKey = `${collectionPrefix}/${docPrefix}/${upload.filename}`
-
-        const { $metadata } = await client.send(
-          new AWS.HeadObjectCommand({ Bucket: TEST_BUCKET, Key: expectedKey }),
-        )
-
-        expect($metadata.httpStatusCode).toBe(200)
-
-        // URL should include the doc prefix as query param
-        expect(upload.url).toEqual(
-          `/api/${mediaWithCompositePrefixesSlug}/file/${String(upload.filename)}?prefix=${docPrefix}`,
-        )
-      })
-
-      it('can upload with composite prefixes (collection prefix only)', async () => {
-        const upload = await payload.create({
-          collection: mediaWithCompositePrefixesSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
+          expect(fileCountAfter).toBe(fileCountBefore)
         })
 
-        expect(upload.id).toBeTruthy()
-
-        // With no doc prefix, file should be at collectionPrefix/filename
-        const expectedKey = `${collectionPrefix}/${upload.filename}`
-
-        const { $metadata } = await client.send(
-          new AWS.HeadObjectCommand({ Bucket: TEST_BUCKET, Key: expectedKey }),
-        )
-
-        expect($metadata.httpStatusCode).toBe(200)
-      })
-
-      it('should not upload to S3 when mimeType validation fails', async () => {
-        // Get initial bucket contents
-        const listBefore = await client.send(new AWS.ListObjectsV2Command({ Bucket: TEST_BUCKET }))
-        const fileCountBefore = listBefore.Contents?.length || 0
-
-        // Try to upload a JSON file to a collection that only accepts PNG
-        await expect(
-          payload.create({
+        it('should upload to S3 when mimeType validation passes', async () => {
+          // Upload a valid PNG file
+          const upload = await payload.create({
             collection: restrictedMediaSlug,
             data: {},
-            filePath: path.resolve(dirname, './test.json'),
-          }),
-        ).rejects.toThrow()
+            filePath: path.resolve(dirname, './image.png'),
+          })
 
-        // Verify no new files were uploaded to S3
-        const listAfter = await client.send(new AWS.ListObjectsV2Command({ Bucket: TEST_BUCKET }))
-        const fileCountAfter = listAfter.Contents?.length || 0
+          expect(upload.id).toBeTruthy()
 
-        expect(fileCountAfter).toBe(fileCountBefore)
+          const filename = upload.filename
+          if (filename == null || filename === '') {
+            throw new Error('expected filename after upload')
+          }
+
+          // Verify the file was uploaded to S3
+          const { $metadata } = await client.send(
+            new AWS.HeadObjectCommand({
+              Bucket: TEST_BUCKET,
+              Key: filename,
+            }),
+          )
+
+          expect($metadata.httpStatusCode).toBe(200)
+        })
+
+        it('should store correct URLs for sized images', async () => {
+          const upload = await payload.create({
+            collection: mediaSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
+
+          const apiResponse = await payload.findByID({
+            id: upload.id,
+            collection: mediaSlug,
+          })
+          expect(apiResponse.sizes).toBeTruthy()
+
+          const apiSizeKeys = Object.keys(apiResponse.sizes || {})
+          for (const sizeKey of apiSizeKeys) {
+            const size = apiResponse.sizes?.[sizeKey as keyof typeof apiResponse.sizes]
+            if (!size) {
+              continue
+            }
+
+            expect(size.url).toEqual(`/api/${mediaSlug}/file/${size.filename}`)
+          }
+
+          const rawDbData = await payload.db.findOne({
+            collection: mediaSlug,
+            where: { id: { equals: upload.id } },
+          })
+          expect(rawDbData).toBeTruthy()
+
+          const dbRecord = rawDbData as unknown as {
+            filename: string
+            sizes: Record<string, { filename: string; url: string }>
+            url: string
+          }
+          type SizeData = { filename: string; url: string }
+
+          const sizeKeys = Object.keys(dbRecord.sizes)
+          expect(sizeKeys.length).toBeGreaterThan(0)
+
+          for (const sizeKey of sizeKeys) {
+            const size: SizeData = dbRecord.sizes[sizeKey] as SizeData
+            expect(size.url).not.toEqual(`/api/${mediaSlug}/file/${dbRecord.filename}`)
+            expect(size.url).toEqual(`/api/${mediaSlug}/file/${size.filename}`)
+          }
+        })
+
+        it('should handle collections without imageSizes correctly', async () => {
+          const upload = await payload.create({
+            collection: mediaWithPrefixSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
+
+          expect(upload.filename).toBeTruthy()
+          expect(upload.url).toEqual(
+            `/api/${mediaWithPrefixSlug}/file/${upload.filename}?prefix=${prefix}`,
+          )
+          expect((upload as any).sizes).toBeFalsy()
+
+          const rawDbData = await payload.db.findOne({
+            collection: mediaWithPrefixSlug,
+            where: { id: { equals: upload.id } },
+          })
+
+          expect(rawDbData).toBeTruthy()
+
+          const dbRecord = rawDbData as unknown as {
+            filename: string
+            url: string
+          }
+
+          expect(dbRecord.filename).toEqual(upload.filename)
+          expect(dbRecord.url).toEqual(`/api/${mediaWithPrefixSlug}/file/${upload.filename}`)
+          expect((rawDbData as any)?.sizes).toBeFalsy()
+        })
+
+        it('should use custom generateFileURL in beforeChange when disablePayloadAccessControl is true', async () => {
+          // This test verifies that custom generateFileURL is used in beforeChange hook
+          // when disablePayloadAccessControl: true, preventing "undefined" from appearing in URLs
+          const upload = await payload.create({
+            collection: mediaWithCustomURLSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
+
+          expect(upload.id).toBeTruthy()
+          expect(upload.filename).toBeTruthy()
+
+          // Get the raw DB data
+          const rawDbData = await payload.db.findOne({
+            collection: mediaWithCustomURLSlug,
+            where: { id: { equals: upload.id } },
+          })
+
+          const dbRecord = rawDbData as unknown as {
+            filename: string
+            prefix: string
+            url: string
+          }
+
+          // The custom generateFileURL should be used, resulting in test-cdn.example.com URL
+          expect(dbRecord.url).toContain('test-cdn.example.com')
+          expect(dbRecord.url).toContain(prefix)
+          expect(dbRecord.url).toContain(encodeURIComponent(dbRecord.filename))
+          expect(dbRecord.url).toMatch(/^https:\/\//)
+
+          // Verify the API response also returns the custom URL
+          const apiResponse = await payload.findByID({
+            id: upload.id,
+            collection: mediaWithCustomURLSlug,
+          })
+
+          expect(apiResponse.url).toContain('test-cdn.example.com')
+          expect(apiResponse.url).toContain(prefix)
+        })
+
+        it('should use custom generateFileURL even without disablePayloadAccessControl', async () => {
+          const upload = await payload.create({
+            collection: mediaWithGenerateFileURLSlug,
+            data: {},
+            filePath: path.resolve(dirname, '../uploads/image.png'),
+          })
+
+          expect(upload.id).toBeTruthy()
+          expect(upload.filename).toBeTruthy()
+
+          const rawDbData = await payload.db.findOne({
+            collection: mediaWithGenerateFileURLSlug,
+            where: { id: { equals: upload.id } },
+          })
+
+          const dbRecord = rawDbData as unknown as {
+            filename: string
+            prefix: string
+            url: string
+          }
+
+          // The URL should be the CDN URL, not a relative path
+          expect(dbRecord.url).toContain('cdn-proxied.example.com')
+          expect(dbRecord.url).toContain(prefix)
+          expect(dbRecord.url).toContain(encodeURIComponent(dbRecord.filename))
+          expect(dbRecord.url).toMatch(/^https:\/\//)
+
+          const apiResponse = await payload.findByID({
+            id: upload.id,
+            collection: mediaWithGenerateFileURLSlug,
+          })
+
+          expect(apiResponse.url).toContain('cdn-proxied.example.com')
+          expect(apiResponse.url).toContain(prefix)
+        })
+      })
+    })
+
+    describe('External data persistence', () => {
+      const createdIDs: (number | string)[] = []
+
+      afterEach(async () => {
+        for (const id of createdIDs) {
+          try {
+            await payload.delete({ id, collection: testMetadataSlug })
+          } catch (e) {
+            // Ignore
+          }
+        }
+        createdIDs.length = 0
       })
 
-      it('should upload to S3 when mimeType validation passes', async () => {
-        // Upload a valid PNG file
+      it('should automatically persist metadata returned by custom adapters', async () => {
         const upload = await payload.create({
-          collection: restrictedMediaSlug,
-          data: {},
+          collection: testMetadataSlug,
+          data: {
+            testNote: 'Testing automatic metadata persistence',
+          },
+          filePath: path.resolve(dirname, '../uploads/image.png'),
+        })
+
+        createdIDs.push(upload.id)
+
+        expect(upload.id).toBeTruthy()
+        expect(upload.filename).toBeTruthy()
+        expect(upload.testNote).toBe('Testing automatic metadata persistence')
+
+        // Our afterChange hook should automatically persist whatever metadata the adapter returns
+        expect(upload.customStorageId).toBeTruthy()
+        expect(upload.customStorageId).toContain('storage-')
+        expect(upload.uploadTimestamp).toBeTruthy()
+        expect(upload.storageProvider).toBe('test-adapter')
+        expect(upload.bucketName).toBe('test-bucket')
+        expect(upload.objectKey).toBe(upload.filename)
+        expect(upload.processingStatus).toBe('completed')
+        expect(upload.uploadVersion).toBe('1.0.0')
+
+        console.log('Test adapter metadata automatically persisted:', {
+          bucketName: upload.bucketName,
+          customStorageId: upload.customStorageId,
+          objectKey: upload.objectKey,
+          processingStatus: upload.processingStatus,
+          storageProvider: upload.storageProvider,
+          uploadTimestamp: upload.uploadTimestamp,
+          uploadVersion: upload.uploadVersion,
+        })
+      })
+
+      it('should persist metadata on update operations', async () => {
+        const upload = await payload.create({
+          collection: testMetadataSlug,
+          data: {
+            testNote: 'Testing update metadata persistence',
+          },
+          filePath: path.resolve(dirname, '../uploads/image.png'),
+        })
+
+        createdIDs.push(upload.id)
+
+        const initialStorageId = upload.customStorageId
+        const initialTimestamp = upload.uploadTimestamp
+
+        const updatedUpload = await payload.update({
+          id: upload.id,
+          collection: testMetadataSlug,
+          data: {
+            testNote: 'Updated test note',
+          },
           filePath: path.resolve(dirname, './image.png'),
         })
 
-        expect(upload.id).toBeTruthy()
+        expect(updatedUpload.testNote).toBe('Updated test note')
 
-        // Verify the file was uploaded to S3
-        const { $metadata } = await client.send(
-          new AWS.HeadObjectCommand({
-            Bucket: TEST_BUCKET,
-            Key: upload.filename,
-          }),
-        )
+        // Test that metadata persistence works on updates too
+        expect(updatedUpload.customStorageId).toBeTruthy()
+        expect(updatedUpload.uploadTimestamp).toBeTruthy()
+        expect(updatedUpload.storageProvider).toBe('test-adapter')
+        expect(updatedUpload.bucketName).toBe('test-bucket')
+        expect(updatedUpload.objectKey).toBe(updatedUpload.filename)
+        expect(updatedUpload.processingStatus).toBe('completed')
+        expect(updatedUpload.uploadVersion).toBe('1.0.0')
 
-        expect($metadata.httpStatusCode).toBe(200)
-      })
+        const filenamesAreDifferent = upload.filename !== updatedUpload.filename
+        const storageIdsAreDifferent = updatedUpload.customStorageId !== initialStorageId
+        const timestampsAreDifferent = updatedUpload.uploadTimestamp !== initialTimestamp
 
-      it('should store correct URLs for sized images', async () => {
-        const upload = await payload.create({
-          collection: mediaSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
+        // If filename changed, storage ID and timestamp should also change (new upload)
+        expect(filenamesAreDifferent).toBe(storageIdsAreDifferent)
+        expect(filenamesAreDifferent).toBe(timestampsAreDifferent)
+
+        console.log('Update test adapter metadata persistence:', {
+          filenameChanged: filenamesAreDifferent,
+          newStorageId: updatedUpload.customStorageId,
+          storageIdChanged: storageIdsAreDifferent,
+          timestampChanged: timestampsAreDifferent,
         })
-
-        const apiResponse = await payload.findByID({
-          collection: mediaSlug,
-          id: upload.id,
-        })
-        expect(apiResponse.sizes).toBeTruthy()
-
-        const apiSizeKeys = Object.keys(apiResponse.sizes || {})
-        for (const sizeKey of apiSizeKeys) {
-          const size = apiResponse.sizes?.[sizeKey as keyof typeof apiResponse.sizes]
-          if (!size) {
-            continue
-          }
-
-          expect(size.url).toEqual(`/api/${mediaSlug}/file/${size.filename}`)
-        }
-
-        const rawDbData = await payload.db.findOne({
-          collection: mediaSlug,
-          where: { id: { equals: upload.id } },
-        })
-        expect(rawDbData).toBeTruthy()
-
-        const dbRecord = rawDbData as unknown as {
-          filename: string
-          sizes: Record<string, { filename: string; url: string }>
-          url: string
-        }
-        type SizeData = { filename: string; url: string }
-
-        const sizeKeys = Object.keys(dbRecord.sizes)
-        expect(sizeKeys.length).toBeGreaterThan(0)
-
-        for (const sizeKey of sizeKeys) {
-          const size: SizeData = dbRecord.sizes[sizeKey] as SizeData
-          expect(size.url).not.toEqual(`/api/${mediaSlug}/file/${dbRecord.filename}`)
-          expect(size.url).toEqual(`/api/${mediaSlug}/file/${size.filename}`)
-        }
-      })
-
-      it('should handle collections without imageSizes correctly', async () => {
-        const upload = await payload.create({
-          collection: mediaWithPrefixSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
-        })
-
-        expect(upload.filename).toBeTruthy()
-        expect(upload.url).toEqual(
-          `/api/${mediaWithPrefixSlug}/file/${upload.filename}?prefix=${prefix}`,
-        )
-        expect((upload as any).sizes).toBeFalsy()
-
-        const rawDbData = await payload.db.findOne({
-          collection: mediaWithPrefixSlug,
-          where: { id: { equals: upload.id } },
-        })
-
-        expect(rawDbData).toBeTruthy()
-
-        const dbRecord = rawDbData as unknown as {
-          filename: string
-          url: string
-        }
-
-        expect(dbRecord.filename).toEqual(upload.filename)
-        expect(dbRecord.url).toEqual(`/api/${mediaWithPrefixSlug}/file/${upload.filename}`)
-        expect((rawDbData as any)?.sizes).toBeFalsy()
-      })
-
-      it('should use custom generateFileURL in beforeChange when disablePayloadAccessControl is true', async () => {
-        // This test verifies that custom generateFileURL is used in beforeChange hook
-        // when disablePayloadAccessControl: true, preventing "undefined" from appearing in URLs
-        const upload = await payload.create({
-          collection: mediaWithCustomURLSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
-        })
-
-        expect(upload.id).toBeTruthy()
-        expect(upload.filename).toBeTruthy()
-
-        // Get the raw DB data
-        const rawDbData = await payload.db.findOne({
-          collection: mediaWithCustomURLSlug,
-          where: { id: { equals: upload.id } },
-        })
-
-        const dbRecord = rawDbData as unknown as {
-          filename: string
-          prefix: string
-          url: string
-        }
-
-        // The custom generateFileURL should be used, resulting in test-cdn.example.com URL
-        expect(dbRecord.url).toContain('test-cdn.example.com')
-        expect(dbRecord.url).toContain(prefix)
-        expect(dbRecord.url).toContain(encodeURIComponent(dbRecord.filename))
-        expect(dbRecord.url).toMatch(/^https:\/\//)
-
-        // Verify the API response also returns the custom URL
-        const apiResponse = await payload.findByID({
-          collection: mediaWithCustomURLSlug,
-          id: upload.id,
-        })
-
-        expect(apiResponse.url).toContain('test-cdn.example.com')
-        expect(apiResponse.url).toContain(prefix)
-      })
-
-      it('should use custom generateFileURL even without disablePayloadAccessControl', async () => {
-        const upload = await payload.create({
-          collection: mediaWithGenerateFileURLSlug,
-          data: {},
-          filePath: path.resolve(dirname, '../uploads/image.png'),
-        })
-
-        expect(upload.id).toBeTruthy()
-        expect(upload.filename).toBeTruthy()
-
-        const rawDbData = await payload.db.findOne({
-          collection: mediaWithGenerateFileURLSlug,
-          where: { id: { equals: upload.id } },
-        })
-
-        const dbRecord = rawDbData as unknown as {
-          filename: string
-          prefix: string
-          url: string
-        }
-
-        // The URL should be the CDN URL, not a relative path
-        expect(dbRecord.url).toContain('cdn-proxied.example.com')
-        expect(dbRecord.url).toContain(prefix)
-        expect(dbRecord.url).toContain(encodeURIComponent(dbRecord.filename))
-        expect(dbRecord.url).toMatch(/^https:\/\//)
-
-        const apiResponse = await payload.findByID({
-          collection: mediaWithGenerateFileURLSlug,
-          id: upload.id,
-        })
-
-        expect(apiResponse.url).toContain('cdn-proxied.example.com')
-        expect(apiResponse.url).toContain(prefix)
-      })
-    })
-  })
-
-  describe('External data persistence', () => {
-    const createdIDs: (number | string)[] = []
-
-    afterEach(async () => {
-      for (const id of createdIDs) {
-        try {
-          await payload.delete({ collection: testMetadataSlug, id })
-        } catch (e) {
-          // Ignore
-        }
-      }
-      createdIDs.length = 0
-    })
-
-    it('should automatically persist metadata returned by custom adapters', async () => {
-      const upload = await payload.create({
-        collection: testMetadataSlug,
-        data: {
-          testNote: 'Testing automatic metadata persistence',
-        },
-        filePath: path.resolve(dirname, '../uploads/image.png'),
-      })
-
-      createdIDs.push(upload.id)
-
-      expect(upload.id).toBeTruthy()
-      expect(upload.filename).toBeTruthy()
-      expect(upload.testNote).toBe('Testing automatic metadata persistence')
-
-      // Our afterChange hook should automatically persist whatever metadata the adapter returns
-      expect(upload.customStorageId).toBeTruthy()
-      expect(upload.customStorageId).toContain('storage-')
-      expect(upload.uploadTimestamp).toBeTruthy()
-      expect(upload.storageProvider).toBe('test-adapter')
-      expect(upload.bucketName).toBe('test-bucket')
-      expect(upload.objectKey).toBe(upload.filename)
-      expect(upload.processingStatus).toBe('completed')
-      expect(upload.uploadVersion).toBe('1.0.0')
-
-      console.log('Test adapter metadata automatically persisted:', {
-        customStorageId: upload.customStorageId,
-        uploadTimestamp: upload.uploadTimestamp,
-        storageProvider: upload.storageProvider,
-        bucketName: upload.bucketName,
-        objectKey: upload.objectKey,
-        processingStatus: upload.processingStatus,
-        uploadVersion: upload.uploadVersion,
       })
     })
 
-    it('should persist metadata on update operations', async () => {
-      const upload = await payload.create({
-        collection: testMetadataSlug,
-        data: {
-          testNote: 'Testing update metadata persistence',
-        },
-        filePath: path.resolve(dirname, '../uploads/image.png'),
-      })
-
-      createdIDs.push(upload.id)
-
-      const initialStorageId = upload.customStorageId
-      const initialTimestamp = upload.uploadTimestamp
-
-      const updatedUpload = await payload.update({
-        collection: testMetadataSlug,
-        id: upload.id,
-        data: {
-          testNote: 'Updated test note',
-        },
-        filePath: path.resolve(dirname, './image.png'),
-      })
-
-      expect(updatedUpload.testNote).toBe('Updated test note')
-
-      // Test that metadata persistence works on updates too
-      expect(updatedUpload.customStorageId).toBeTruthy()
-      expect(updatedUpload.uploadTimestamp).toBeTruthy()
-      expect(updatedUpload.storageProvider).toBe('test-adapter')
-      expect(updatedUpload.bucketName).toBe('test-bucket')
-      expect(updatedUpload.objectKey).toBe(updatedUpload.filename)
-      expect(updatedUpload.processingStatus).toBe('completed')
-      expect(updatedUpload.uploadVersion).toBe('1.0.0')
-
-      const filenamesAreDifferent = upload.filename !== updatedUpload.filename
-      const storageIdsAreDifferent = updatedUpload.customStorageId !== initialStorageId
-      const timestampsAreDifferent = updatedUpload.uploadTimestamp !== initialTimestamp
-
-      // If filename changed, storage ID and timestamp should also change (new upload)
-      expect(filenamesAreDifferent).toBe(storageIdsAreDifferent)
-      expect(filenamesAreDifferent).toBe(timestampsAreDifferent)
-
-      console.log('Update test adapter metadata persistence:', {
-        filenameChanged: filenamesAreDifferent,
-        storageIdChanged: storageIdsAreDifferent,
-        timestampChanged: timestampsAreDifferent,
-        newStorageId: updatedUpload.customStorageId,
-      })
-    })
-  })
-
-  describe('Azure', () => {
-    it.todo('can upload')
-  })
-
-  describe('GCS', () => {
-    it.todo('can upload')
-  })
-
-  describe('R2', () => {
-    it.todo('can upload')
-  })
-
-  async function verifyUploads({
-    collectionSlug,
-    uploadId,
-    prefix = '',
-  }: {
-    collectionSlug: keyof Config['collections']
-    prefix?: string
-    uploadId: number | string
-  }) {
-    const uploadData = (await payload.findByID({
-      collection: collectionSlug,
-      id: uploadId,
-    })) as unknown as { filename: string; sizes: Record<string, { filename: string }> }
-
-    const fileKeys = Object.keys(uploadData.sizes || {}).map((key) => {
-      const rawFilename = uploadData.sizes[key].filename
-      return prefix ? `${prefix}/${rawFilename}` : rawFilename
+    describe('Azure', () => {
+      it.todo('can upload')
     })
 
-    fileKeys.push(`${prefix ? `${prefix}/` : ''}${uploadData.filename}`)
-    try {
-      for (const key of fileKeys) {
-        const { $metadata } = await client.send(
-          new AWS.HeadObjectCommand({ Bucket: TEST_BUCKET, Key: key }),
-        )
+    describe('GCS', () => {
+      it.todo('can upload')
+    })
 
-        if ($metadata.httpStatusCode !== 200) {
-          console.error('Error verifying uploads', key, $metadata)
-          throw new Error(`Error verifying uploads: ${key}, ${$metadata.httpStatusCode}`)
-        }
-
-        // Verify each size was properly uploaded
-        expect($metadata.httpStatusCode).toBe(200)
-      }
-    } catch (error: unknown) {
-      console.error('Error verifying uploads:', fileKeys, error)
-      throw error
-    }
-  }
+    describe('R2', () => {
+      it.todo('can upload')
+    })
+  })
 })
